+++ frontend/src/App.tsx
@@ -3,9 +3,9 @@ import {useEffect, useState} from "react";
 import {Assignment, Course, CourseDto, LessonDto, NewCourseDto} from "./types/types.ts";
 import axios, {AxiosResponse} from 'axios';
 import {Route, Routes, useNavigate} from "react-router-dom";
-import CoursePage from "./components/CoursePage.tsx";
+import CoursePage from "./pages/CoursePage.tsx";
 import CourseCreator from "./components/CourseCreator.tsx";
-import Dashboard from "./components/Dashboard.tsx";
+import Dashboard from "./pages/Dashboard.tsx";
 import {formatDate} from "./utils/formatDate.ts";
 import CourseLessonOverview from "./components/CourseLessonOverview.tsx";
 import CourseAssignmentOverview from "./components/CourseAssignmentOverview.tsx";
